Fixed FileInput Errors
I adjusted the code in the FileInput class to work with the Sorter class methods, so now command line tests can begin. However, there is a feature I would like to implement in the code that enters a new line after an amount of words are printed out which explains why there is hard code in the Sorter class and why the for-each loop was changed to a for loop. Try to implement this feature as it would help readability when long file inputs are passed through.