File Tests
Input file is read, however, the words are not being read correctly - my guess is that the words with punctuation or punctuation alone is messing up the output file. Output file works for the English -> Morse but not for Morse -> English as nothing is written into the file. I tried manually translating the files and although it did translate, the process has revealed another input error where multiple inputs are needlessly taken - I assume the long text entered has to do with this.